refactor: removing old component and styles, renaming component
