--- app/routes/account.$accountId.$appId.notifications/components/NotificationsAlertForm/styles.css
@@ -1,54 +1,0 @@
-.pokt-network-notifications-activate-alerts {
-  width: 100%;
-}
-
-.pokt-network-notifications-alerts .pokt-card {
-  padding: 24px;
-}
-
-.pokt-network-notifications-alerts {
-  margin: 32px 0;
-}
-
-.pokt-network-notifications-alert {
-  position: relative;
-  width: 100%;
-  padding-left: 24px;
-  margin-bottom: 16px;
-}
-
-.pokt-network-notifications-alert-border {
-  position: absolute;
-  left: 0px;
-  top: 0px;
-  width: 8px;
-  height: 100%;
-  border-radius: 8px 0px 0px 8px;
-}
-
-.pokt-network-notifications-alert-btn-container {
-  width: 100%;
-  display: flex;
-  justify-content: flex-end;
-}
-
-.pokt-network-notifications-alert-border-positive {
-  background-color: green;
-}
-
-.pokt-network-notifications-alert-border-yellow {
-  background-color: yellow;
-}
-
-.pokt-network-notifications-alert-border-warning {
-  background-color: orange;
-}
-
-.pokt-network-notifications-alert-border-negative {
-  background-color: red;
-}
-
-.pokt-network-notifications-alert-description {
-  display: flex;
-  justify-content: space-between;
-}
