+++ app/routes/account.$accountId.$appId.notifications/route.tsx
@@ -4,7 +4,7 @@ import { useEffect } from "react"
 import invariant from "tiny-invariant"
 import { AppIdOutletContext } from "../account.$accountId.$appId/route"
 import { initPortalClient } from "~/models/portal/portal.server"
-import AppNotificationsAlert from "~/routes/account.$accountId.$appId.notifications/components/NotificationsAlertForm"
+import AppNotificationsAlert from "~/routes/account.$accountId.$appId.notifications/components/AppNotificationsAlert"
 import { AmplitudeEvents, trackEvent } from "~/utils/analytics"
 import { requireUser } from "~/utils/session.server"
